Delete measumente legacy is unused, move function get_date to date utils
