+++ backend-refactor/app/factory.py
@@ -1,13 +1,12 @@
 """Application factory and configuration."""
 from fastapi import FastAPI
 from fastapi.staticfiles import StaticFiles
-from fastapi.templating import Jinja2Templates
+#from fastapi.templating import Jinja2Templates
 
 from app.api.v1.endpoints import assets as assets_v1
 from app.api.v2.routes import measurements_router
 from app.core.config import get_settings
 from app.core.settings import AppSettings
-from app.services import measurement_legacy
 
 def create_app() -> FastAPI:
     """Create and configure the FastAPI application."""
@@ -23,7 +22,6 @@ def create_app() -> FastAPI:
     app.include_router(measurements_router.router, prefix="/api/v1")
     app.include_router(assets_v1.router, tags=["assets"])
     app.include_router(measurements_router.router, tags=["measurement"])
-    app.include_router(measurement_legacy.router, tags=["measurements"])
     #app.include_router(health_check.router, prefix="/health")
 
     # mount necessary folders
