+++ backend-refactor/app/services/__init__.py
@@ -1,9 +1,8 @@
 """Services package."""
 from app.services.asset_service import AssetService
 from app.services.measurement_svc import MeasurementService, get_measurements_for_signals
-from app.services import measurement_legacy
 
-__all__ = ["measurement_legacy",
+__all__ = [
     "AssetService",
     "MeasurementService", "get_measurements_for_signals"
 ]
