+++ backend-refactor/app/utils/__init__.py
@@ -1,12 +1,12 @@
 """Utilities package."""
 from app.utils.asset_helper import format_asset_response, FormatAssetResponse, transform_asset, process_asset_data
 from app.utils.helpers import validate_data, check_data, is_valid, format_response
-from app.utils.date_utils import parse_date, ParseDate, validate_date_range, check_date_range, is_valid_date_range
+from app.utils.date_utils import parse_date, ParseDate, validate_date_range, check_date_range, is_valid_date_range, get_date
 from app.utils.measurement_utils import filter_measurements_by_date, FilterMeasurementsByDate, format_measurement
 
 __all__ = [
     "format_asset_response", "FormatAssetResponse", "transform_asset", "process_asset_data",
     "validate_data", "check_data", "is_valid", "format_response",
-    "parse_date", "ParseDate", "validate_date_range", "check_date_range", "is_valid_date_range",
-    "filter_measurements_by_date", "FilterMeasurementsByDate", "format_measurement"
+    "parse_date", "ParseDate", "validate_date_range", "check_date_range", "is_valid_date_range", "get_date",
+    "filter_measurements_by_date", "FilterMeasurementsByDate", "format_measurement", 
 ]
