+++ backend-refactor/app/utils/date_utils.py
@@ -1,4 +1,9 @@
 """Date and time utilities."""
+import csv
+from typing import List, Dict
+import datetime
+import os
+
 from datetime import datetime
 from typing import Optional
 
@@ -23,3 +28,27 @@ def check_date_range(start: datetime, end: datetime) -> bool:
 def is_valid_date_range(from_dt: datetime, to_dt: datetime) -> bool:
     """Yet another date range validator."""
     return from_dt < to_dt
+
+def get_date() -> List[Dict]:
+    measurements = []
+    csv_path = os.path.join(os.path.dirname(__file__), "..", "data", "measurements.csv")
+
+    with open(csv_path, 'r', encoding='utf-8-sig') as f:
+        reader = csv.DictReader(f, delimiter='|')
+        for i, row in enumerate(reader):
+            if i >= 500:
+                break
+            measurements.append(row)
+
+    rows = []
+    for row in measurements:
+        rows.append({
+            "signal_id": row.get("SignalId"),
+            "timestamp": row.get("Ts"),
+            "value": row.get("MeasurementValue"),
+            "unit": row.get("Unit")
+        })
+
+    print(measurements[0])
+    return rows 
+    #return datetime.datetime.now().isoformat()
